remove setup.d scripts' need for boilerplate header
it has been annoying to have to have the `source`, REPO_DIR, and
CONFIG_DIR lines at the beginning of every setup.d script. It is a
barrier to adding more scripts, and is brittle to changes, making
changes to that header difficult to do (because all scripts have to be
updated), and easy to mess up. This commit sets the environment
variables when calling the setup.d scripts, so that they don't have to
do it themselves, and moves the helper functions into dedicated scripts,
accessible in the PATH, so setup.d scripts don't need to `source`
setup.sh anymore.