moving kitty-themes to a more properly named dir
I named the directory `themes` originally, to keep it cleaner, but I
didn't appreciate that the kitty-themes repo had a lot more in it than
just themes, and had its own `themes` directory. So I ended up not
liking having to write `themes/themes` in my kitty config, and I feel
like `kitty-themes` better expresses that this directory is not just a
folder of themes, but the kitty-themes repo itself.